// types/supabase.ts
export interface Database {
	graphql_public: {
		Tables: {
			[_ in never]: never;
		};
		Views: {
			[_ in never]: never;
		};
		Functions: {
			graphql: {
				Args: {
					operationName?: string;
					query?: string;
					variables?: Json;
					extensions?: Json;
				};
				Returns: Json;
			};
		};
		Enums: {
			[_ in never]: never;
		};
		CompositeTypes: {
			[_ in never]: never;
		};
	};
	public: {
		Tables: {
			markdowns: {
				Row: {
					content: string | null;
					created_at: string;
					id: string;
					title: string;
					user_id: string;
				};
				Insert: {
					content?: string | null;
					created_at?: string;
					id?: string;
					title: string;
					user_id: string;
				};
				Update: {
					content?: string | null;
					created_at?: string;
					id?: string;
					title?: string;
					user_id?: string;
				};
				Relationships: [
					{
						foreignKeyName: "markdowns_user_id_fkey";
						columns: ["user_id"];
						isOneToOne: false;
						referencedRelation: "users";
						referencedColumns: ["id"];
					}
				];
			};
			waitlist: {
				Row: {
					created_at: string;
					email: string;
					id: number;
				};
				Insert: {
					created_at?: string;
					email: string;
					id?: number;
				};
				Update: {
					created_at?: string;
					email?: string;
					id?: number;
				};
				Relationships: [];
			};
		};
		Views: {
			[_ in never]: never;
		};
		Functions: {
			[_ in never]: never;
		};
		Enums: {
			[_ in never]: never;
		};
		CompositeTypes: {
			[_ in never]: never;
		};
	};
// ...
}